TDD: tests that fail due to #45
