+++ testapp/models.py
@@ -44,7 +44,23 @@ class UUIDModel(models.Model):
 
 
 class TestUniqueNullableModel(models.Model):
+    # Issue #38:
     # This field started off as unique=True *and* null=True so it is implemented with a filtered unique index
     # Then it is made non-nullable by a subsequent migration, to check this is correctly handled (the index
     # should be dropped, then a normal unique constraint should be added, now that the column is not nullable)
     test_field = models.CharField(max_length=100, unique=True)
+
+    # Issue #45 (case 1)
+    # Field used for testing changing the 'type' of a field that's both unique & nullable
+    x = models.CharField(max_length=11, null=True, unique=True)
+
+
+class TestNullableUniqueTogetherModel(models.Model):
+    class Meta:
+        unique_together = (('a', 'b', 'c'),)
+
+    # Issue #45 (case 2)
+    # Fields used for testing changing the 'type of a field that is in a `unique_together`
+    a = models.CharField(max_length=51, null=True)
+    b = models.CharField(max_length=50)
+    c = models.CharField(max_length=50)
